memory error fix
memory error fix

+++ src/pandemia/C/src/default_model_functions.c
@@ -663,7 +663,7 @@ void dynamics_movement
     // Determine new locations, for those agents starting a new activity
     for(int n=0; n<N; n++){
         if(current_region[n] == id){
-            int new_activity = weekly_routines[(n * ticks_in_week) + t_next];
+            uint8_t new_activity = weekly_routines[(n * ticks_in_week) + t_next];
             if(new_activity != weekly_routines[(n * ticks_in_week) + t_now]){
                 // If the agent is not currently wearing a facemask but should be, put it on
                 if(current_facemask[n] == 0 && facemask_intervention == 1 &&
